% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Expit}
\usage{
expit(x)
}
\arguments{
\item{x}{A real number \eqn{x}}
}
\value{
\eqn{1/(1 + exp(-x))}
}
\description{
Inverse of the \code{\link{logit}} function, defined as
\eqn{expit(x) = 1/(1 + \exp(-x))}
}
\examples{
p <- 0.3

x <- logit(p)
expit(x)


}
