% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eta_pt.R
\name{get_eta_pt}
\alias{get_eta_pt}
\title{Select 'eta' hyper-parameter for Markov random field prior}
\usage{
get_eta_pt(
  y,
  X,
  psi,
  R = matrix(0, ncol(X), ncol(X)),
  hyper_params = NULL,
  burnin,
  draws,
  thinning = 5,
  refine_betas = TRUE,
  model_skewness = TRUE,
  adaptive = FALSE,
  adapt_prop = 0.25,
  theta2 = 2,
  vs = TRUE,
  pmax = ncol(X),
  pmax_draws = 10,
  eta_tol = 0.1,
  percentile_check = 0.1,
  quantile_change_max = 0.05
)
}
\arguments{
\item{y}{Numeric vector containing observations of the outcome variable.}

\item{X}{Numeric matrix containing observations of the predictor variables.}

\item{psi}{Numeric detection limit (optional). If specified, y values below this
value will be treated as left-censored values from the skewed normal distribution.}

\item{R}{Relationship matrix for Markov Random Field variable selection prior (optional).
If specified, must be a symmetric non-negative matrix with diagonal zero and
dimension \code{ncol(X)}. If not specified, the
variable selection prior will be independent across predictors. See details.}

\item{hyper_params}{List of hyper-parameters for the MCMC. See details.}

\item{burnin}{Number of burnin draws for the MCMC.}

\item{draws}{Number of post-burnin draws for the MCMC.}

\item{thinning}{Thinning parameter for the MCMC. Default is 1, meaning every
post-burnin draw is retained.}

\item{refine_betas}{Logical value indicating whether the beta parameters be updated
separately from the gamma parameters using a Gibbs sampler. Default is TRUE.}

\item{model_skewness}{Logical value indicating whether to use the skew-normal
distribution rather than the normal distribution. Default is TRUE.}

\item{adaptive}{Logical value indicating whether to tune the Metropolis-Hastings
parameter based on samples drawn during the burnin stage of the MCMC. Default
is false because my code for this isn't working.}

\item{adapt_prop}{The proportion of burnin draws used to tune the Metropolis-Hastings
parameter.}

\item{theta2}{Sampling variance of the proposed beta in the Metropolis-Hastings. If adaptive,
this parameter is updated based on the last adapt_prop of the burnin draws.}

\item{vs}{Logical value indicating whether to perform variable selection. Default is true.}

\item{pmax}{Upper limit for the number of selected variables. If more than \code{pmax} variable are selected
in \code{pmax_draws} consecutive draws, the algorithm is terminated early to prevent phase transition. The default
value is \code{ncol(X)}, meaning there is no early stopping.}

\item{pmax_draws}{Used with \code{pmax} to prevent phase transition. Default is \code{10}.}

\item{eta_tol}{Error bound for eta. Default is \code{0.1}.}

\item{percentile_check}{Percentile used to check for phase transition. The default
is \code{0.1}, which means phase transition is determined
by tracking the 10th percentile of the \eqn{\hat{\gamma}} vector.}

\item{quantile_change_max}{Maximum allowable change in the quantile corresponding
to \code{percentile_check}. Default is 0.05. The algorithm identifies phase
transition if the \code{percentile_check} percentile of the \eqn{\hat{\gamma}}
vector is mode than \code{quantile_change_max} greater than the model with \eqn{eta=0}.}
}
\value{
The largest \eqn{eta} value such that the model does not experience
phase transition.
}
\description{
Select 'eta' hyper-parameter for Markov random field prior
}
