% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logit}
\usage{
logit(p)
}
\arguments{
\item{p}{A probability \eqn{p}.}
}
\value{
\eqn{log(p/(1-p))}
}
\description{
For a probability \eqn{p}, returns \eqn{\log(p/(1-p))}.
}
\examples{

p <- 0.34
logit(p)

}
