% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metab_vs_hyperparams.R
\name{metab_vs_hyperparams}
\alias{metab_vs_hyperparams}
\title{Generate hyper-parameters for Bayesian variable selection}
\usage{
metab_vs_hyperparams(
  method = c("default", "tuned"),
  nu2_0 = 10^2,
  nu2_d = 10^2,
  nu2 = rep(2^2, p),
  lambda2 = rep(10^2, s),
  xi_0 = 5,
  sigma2_0 = 4,
  omega = logit(0.05),
  eta = 0,
  rho_0 = 2 * 0.9,
  rho_1 = 2 * 0.1,
  y = NULL,
  psi = NULL,
  X = NULL,
  C = NULL
)
}
\arguments{
\item{method}{Either \code{"default"} (the default, namely) or \code{"tuned"}.
If \code{"default"}, simple uninformative priors are used for all parameters.
If \code{"tuned"}, the priors are chosen empirically based on the observed data,
which must be provided.}

\item{nu2_0}{Prior variance of the intercept, \code{\beta_0\sim\mathcal{N}(0,\nu_0^2)}. T}

\item{nu2_d}{Prior variance of the delta, \code{\delta\sim\mathcal{N}(0,\nu_d^2)}.}

\item{nu2}{Prior variance of the regression coefficients of each \code{\X_{ji}},
\code{\beta_j\sim\mathcal{N}(0,\nu^2)}, \code{j\in 1,\dots,p}.}

\item{lambda2}{Prior variance of the regression coefficients of each \code{\C_{ji}},
\code{\alpha_j\sim\mathcal{N}(0,\lambda_j^2)}, \code{j\in 1,\dots,p}.}

\item{omega}{Hyperparameter of MRF hyperprior that controls overall sparsity}

\item{eta}{Hyperparameter of}

\item{y}{(If \code{method="tuned"}) Numeric vector of metabolite measurements. Values will be
treated as point mass values (PMVs) if they are \code{NA} or if they are below \code{psi}.}

\item{psi}{(If \code{method="tuned"}) Numeric value representing the metabolite's detection limit. All
\code{y} values below \code{psi} will be treated as PMVs and inferred via data
augmentation. The default is \code{min(y, na.rm =TRUE)}, in which case only
\code{NA} values of \code{y} will be treated as PMVs.}

\item{X}{(If \code{method="tuned"}) Numeric matrix of the predictor variables that will
be considered for variable selection. Rows are the subjects and columns
are the variables.}

\item{C}{(If \code{method="tuned"}) Numeric matrix of confounding variables that are excluded
from variable selection (optional). Rows are the the subjects and columns are the variables.
If provided, should NOT include an intercept term.}

\item{R}{(If \code{method="tuned"}) Relationship matrix for Markov Random Field variable selection prior (optional).
If specified, must be a symmetric non-negative matrix with diagonal zero and
dimension \code{ncol(X)}. If not specified, the
variable selection hyperprior will be independent and identically across
predictors, with prior inclusion probability \code{expit(omega)}. See \code{\link[=metab_bvs]{metab_bvs()}}.}
}
\description{
Generate hyper-parameters for Bayesian variable selection
}
