% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolites.R
\docType{data}
\name{metabolites}
\alias{metabolites}
\title{Simulated metabolite data}
\format{
\subsection{\code{metabolites}}{

A list with two elements:
\describe{
\item{y}{Metabolite values for 100 subjects}
\item{X}{Numeric matrix of 10 predictor variables}
\item{C}{Numeric matrix of 3 confounding variables}
\item{psi}{Detection limit of the metabolite. \code{Y} values below
\code{psi} are censored}
\item{R}{Covariate relationship matrix used in Markov Random Field prior}
...
}
}
}
\usage{
metabolites
}
\description{
Metabolite data simulated from a skew-normal mixture model.
}
\keyword{datasets}
