% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_to_variables.R
\name{pip_to_variables}
\alias{pip_to_variables}
\title{Determine PIPs}
\usage{
pip_to_variables(pip, method = c("fdr", "cutoff"), cutoff = 0.5, fdr = 0.05)
}
\arguments{
\item{pip}{The pips}

\item{method}{the method}

\item{cutoff}{cutoff}

\item{fdr}{fdr}
}
\description{
Determine PIPs
}
