% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_variables.R
\name{selected_variables}
\alias{selected_variables}
\title{Identify selected variables from Bayesian model}
\usage{
selected_variables(fit, method = c("fdr", "cutoff"), cutoff = 0.5, fdr = 0.05)
}
\arguments{
\item{fit}{A \code{list} as outputted by \code{metab_vs}.}

\item{method}{Either \code{"cutoff"} when providing a PIP cutoff to determine
significance, or \code{"fdr"} to determine the cutoff automatically using Bayesian FDR rules.
Default is \code{"fdr"}.}

\item{cutoff}{If \code{method = "cutoff"}, the PIP cutoff for declaring
significance. Default is \code{0.5}.}

\item{fdr}{If \code{method = "cutoff"}, the nominal FDR rate. Default is \code{0.05}.}
}
\value{
A list containing selection indicators (\code{variables}) and the
PIP cutoff (\code{cutoff}).
}
\description{
Identify selected variables from Bayesian model
}
