% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_metab.R
\name{standardize_metab}
\alias{standardize_metab}
\title{Standardize metabolites}
\usage{
standardize_metab(Y, psi = apply(Y, 2, min, na.rm = TRUE))
}
\arguments{
\item{Y}{Numeric matrix of metabolite values. Each row is a subject and
each column is a metabolite. Values will be treated as point mass values
(PMVs) if they are \code{NA} or if they are below \code{psi}.}

\item{psi}{Numeric vector of metabolite-specific detection limits, length \code{ncol(Y)}.
Default is \code{apply(Y,2,min,na.rm=TRUE)}.
All \code{Y} values below their corresponding \code{psi} will be treated as
PMVs (in addition to any \code{Y} values that are \code{NA}).}
}
\value{
List of the following components:
\itemize{
\item \code{Y}: A matrix of standardized metabolite values
\item \code{means}: A vector of means used for standardization
\item \code{sds}: A vector of standard deviations used for standardization.
}
}
\description{
Standardize metabolites
}
\details{
Standardizing \code{Y} before analysis is essential for specifying hyperparameters.
However, this step is not trivial since \code{Y} is expected to contain
missing values called point mass values (PMVs) (see \code{\link[=metab_vs]{metab_vs()}}). To resolve
this, we impute \code{Y} using the half-minimum rule, standardize, then
re-introduce the missing values. This approach, though imperfect, is useful
for placing metabolites on a comparable numerical scale.
}
