% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metab_vs.R
\name{metab_vs}
\alias{metab_vs}
\title{Bayesian variable selection for metabolite outcomes}
\usage{
metab_vs(
  y,
  X,
  C = NULL,
  psi,
  R = matrix(0, ncol(X), ncol(X)),
  hyper_params = NULL,
  burnin,
  draws,
  thinning = 1,
  refine_betas = TRUE,
  model_skewness = TRUE,
  type = c("MH", "gibbs"),
  adaptive = FALSE,
  adapt_prop = 0.25,
  theta2 = 2,
  psamp = ceiling(p/5),
  vs = TRUE,
  pmax = ncol(X),
  pmax_draws = 10
)
}
\arguments{
\item{y}{Numeric vector containing observations of the outcome variable.}

\item{X}{Numeric matrix containing observations of the predictor variables. Rpws
are the subjects and columns are the variables.}

\item{C}{Optional numeric matrix containing observations of the covariates /
confounding variables. Rows are the the subjects and columns are the variables.
If provided, should NOT include an intercept term.}

\item{psi}{Numeric detection limit (optional). If specified, y values below this
value will be treated as left-censored values from the skewed normal distribution.}

\item{R}{Relationship matrix for Markov Random Field variable selection prior (optional).
If specified, must be a symmetric non-negative matrix with diagonal zero and
dimension \code{ncol(X)}. If not specified, the
variable selection prior will be independent across predictors. See details.}

\item{hyper_params}{List of hyper-parameters for the MCMC. See details.}

\item{burnin}{Number of burnin draws for the MCMC.}

\item{draws}{Number of post-burnin draws for the MCMC.}

\item{thinning}{Thinning parameter for the MCMC. Default is 1, meaning every
post-burnin draw is retained.}

\item{refine_betas}{Logical value indicating whether the beta parameters be updated
separately from the gamma parameters using a Gibbs sampler. Default is TRUE.}

\item{model_skewness}{Logical value indicating whether to use the skew-normal
distribution rather than the normal distribution. Default is TRUE.}

\item{type}{Sampling scheme used for variable selection. Either \code{"MH"}
for Metropolis-Hastings (the default) or \code{"gibbs"} for Gibbs sampling.}

\item{adaptive}{Logical value indicating whether to tune the Metropolis-Hastings
parameter based on samples drawn during the burnin stage of the MCMC. Default
is false because my code for this isn't working.}

\item{adapt_prop}{The proportion of burnin draws used to tune the Metropolis-Hastings
parameter.}

\item{theta2}{Sampling variance of the proposed beta in the Metropolis-Hastings. If adaptive,
this parameter is updated based on the last adapt_prop of the burnin draws.}

\item{vs}{Logical value indicating whether to perform variable selection. Default is true.}

\item{pmax}{Upper limit for the number of selected variables. If more than \code{pmax} variable are selected
in \code{pmax_draws} consecutive draws, the algorithm is terminated early to prevent phase transition. The default
value is \code{ncol(X)}, meaning there is no early stopping.}

\item{pmax_draws}{Used with \code{pmax} to prevent phase transition. Default is \code{10}.}
}
\value{
List containing the following components:
}
\description{
Bayesian variable selection for metabolite outcomes
}
\examples{
# Generate Data
n <- 50
beta <- c(1,1,0,0)
p <- length(beta)
beta0 <- 0
delta <- 3
sigma <- 1
rho <- 0.9
psi <- 0.5
X <- as.matrix(scale(matrix(rnorm(200), nrow = n, ncol = 4)))
u <- rbinom(n, size = 1, prob = rho)
y <- beta0 + c(X \%*\% beta) + sigma * rnorm(n) + delta * abs(rnorm(n))
out <- metab_vs(
    y = y, X = X, psi = psi,
    burnin = 100,
    draws = 100

  )


}
